objtool/sync-check.sh: set the exit code explicitly
Commit 6ec14aa7a58a "objtool: Silence build output" removed 'cd -' from
the end of sync-check.sh script. As a side-effect, the script now exits
with the exit code of the last executed 'check' command. If any of the
files it should check are missing (which is not an error per se), it will
return 1 and objtool will fail to build.

Explicit 'exit 0' at the end should fix that.

Fixes: 6ec14aa7a58a "objtool: Silence build output"
Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>