Remove RPM-illegal chars from module version
$ modinfo aacraid.ko | grep ^version:
version: 1.2.1[50877]-custom

kmod-deps.sh from rpm-openmandriva-setup is an RPM generator of Provides, it generates:
"kmod(aacraid) = 1.2.1[50877]_custom"
'[' and ']' are illegal characters in RPM

I have no idea how to deal with it in kmod-deps.sh and do not want to drop this/these provide(s),
so let me just patch the module version inside the kernel.

Converting from '1.2.1[50877]-custom' to '1.2.1_50877-custom'
No idea if any external scripts rely on parsing this version, hope that no.

Upstreamized: not needed
Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>