ROSA: ima: allow to off modules signature check dynamically
Allow module.sig_enforce=0 kernel cmdline, not only module.sig_enforce=1
It allows to keep CONFIG_MODULE_SIG_FORCE=y, but disable it when really needed
without recompiling the kernel (it may be impossible, e.g. in certified systems).

GRUB or another bootloader is password-protected when needed,
so I am not afraid much that someone will be able to turn it off when not needed.

ROSA-specific patch.
This violates requirements of "secure boot", but currently we do not have secure boot in ROSA.

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>